refactor: types/types 중복표현 제거를 위한 폴더구조 변경
